// src/app/model/trMember.ts
export class trMember {
    member_id:    any;
    username:     any;
    password:     any;
    profile_name:   any;
    email:      any;
    birthday:    any;
    gender:  any;
    token:    any;
    user_type:   any;
    login_with:    any;
    phone: any;
    active:      any;
    createDate: any;
    createBy:   any;
    updateDate:  any;
    updateBy:    any;
}